+++ src/meltano/api/controllers/orchestrations.py
@@ -9,7 +9,7 @@ from werkzeug.exceptions import Conflict, UnprocessableEntity
 
 from meltano.core.job import JobFinder, State
 from meltano.core.behavior.canonical import Canonical
-from meltano.core.plugin import PluginRef, Profile
+from meltano.core.plugin import PluginRef
 from meltano.core.plugin.error import PluginExecutionError, PluginLacksCapabilityError
 from meltano.core.plugin.settings_service import (
     PluginSettingsService,
@@ -47,18 +47,30 @@ from .upload_helper import InvalidFileTypeError, InvalidFileSizeError, UploadHel
 from .utils import enforce_secure_filename
 
 
-def decorate_profile_config(profile):
-    profile["config"] = freeze_keys(profile.get("config", {}))
+def get_config_with_metadata(settings):
+    config_dict = {}
+    config_metadata = {}
 
-    config_metadata = profile.get("config_metadata", {})
-    for key, metadata in config_metadata.items():
+    config_with_metadata = settings.config_with_metadata(
+        extras=False, redacted=True, session=db.session
+    )
+    for key, metadata in config_with_metadata.items():
+        config_dict[key] = metadata.pop("value")
+
+        metadata.pop("setting", None)
+
         try:
             metadata["source_label"] = metadata["source"].label
             metadata["auto_store_label"] = metadata["auto_store"].label
         except KeyError:
             pass
 
-    profile["config_metadata"] = freeze_keys(config_metadata)
+        config_metadata[key] = metadata
+
+    return {
+        "config": freeze_keys(config_dict),
+        "config_metadata": freeze_keys(config_metadata),
+    }
 
 
 def validate_plugin_config(
@@ -71,7 +83,7 @@ def validate_plugin_config(
         return False
 
     if setting_def.kind == "file" and value and value != "":
-        uploads_directory = project.extract_dir(plugin.full_name)
+        uploads_directory = project.extract_dir(plugin.name)
         resolved_file_path = project.root_dir(value).resolve()
         if not str(resolved_file_path).startswith(str(uploads_directory) + "/"):
             logging.warning(
@@ -259,7 +271,7 @@ def run():
 @block_if_readonly
 def upload_plugin_configuration_file(plugin_ref) -> Response:
     """
-    Endpoint for uploading a file for a specific plugin's configuration profile
+    Endpoint for uploading a file for a specific plugin
     """
 
     file = request.files["file"]
@@ -268,7 +280,7 @@ def upload_plugin_configuration_file(plugin_ref) -> Response:
 
     project = Project.find()
     directory = project.extract_dir(
-        plugin_ref.full_name, ("tmp" if tmp else ""), setting_name
+        plugin_ref.name, ("tmp" if tmp else ""), setting_name
     )
     upload_helper = UploadHelper()
     file_path = upload_helper.upload_file(directory, file)
@@ -282,7 +294,7 @@ def upload_plugin_configuration_file(plugin_ref) -> Response:
 @block_if_readonly
 def delete_plugin_configuration_file(plugin_ref) -> Response:
     """
-    Endpoint for deleting a file for a specific plugin's configuration profile
+    Endpoint for deleting a file for a specific plugin
     """
 
     payload = request.get_json()
@@ -291,7 +303,7 @@ def delete_plugin_configuration_file(plugin_ref) -> Response:
 
     project = Project.find()
     directory = project.extract_dir(
-        plugin_ref.full_name, ("tmp" if tmp else ""), setting_name
+        plugin_ref.name, ("tmp" if tmp else ""), setting_name
     )
     shutil.rmtree(directory)
 
@@ -301,7 +313,7 @@ def delete_plugin_configuration_file(plugin_ref) -> Response:
 @orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration", methods=["GET"])
 def get_plugin_configuration(plugin_ref) -> Response:
     """
-    Endpoint for getting a plugin's configuration profiles
+    Endpoint for getting a plugin's configuration
     """
 
     project = Project.find()
@@ -323,45 +335,13 @@ def get_plugin_configuration(plugin_ref) -> Response:
     except PluginNotFoundError:
         settings_group_validation = []
 
-    profiles = settings.profiles_with_config(redacted=True, session=db.session)
-    for profile in profiles:
-        decorate_profile_config(profile)
-
     return jsonify(
         {
-            "profiles": profiles,
+            **get_config_with_metadata(settings),
             "settings": Canonical.as_canonical(settings.definitions(extras=False)),
             "settings_group_validation": settings_group_validation,
         }
     )
-
-
-@orchestrationsBP.route(
-    "/<plugin_ref:plugin_ref>/configuration/profiles", methods=["POST"]
-)
-@block_if_readonly
-def add_plugin_configuration_profile(plugin_ref) -> Response:
-    """
-    Endpoint for adding a configuration profile to a plugin
-    """
-    payload = request.get_json()
-    project = Project.find()
-    config = ConfigService(project)
-    plugin = config.get_plugin(plugin_ref)
-    settings = PluginSettingsService(project, plugin, config_service=config)
-
-    # create the new profile for this plugin
-    name = payload["name"]
-    profile = plugin.add_profile(slugify(name), label=name)
-
-    config.update_plugin(plugin)
-
-    profile_config = settings.profile_with_config(
-        profile, redacted=True, session=db.session
-    )
-    decorate_profile_config(profile_config)
-
-    return jsonify(profile_config)
 
 
 @orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration", methods=["PUT"])
@@ -375,31 +355,21 @@ def save_plugin_configuration(plugin_ref) -> Response:
     config_service = ConfigService(project)
     plugin = config_service.get_plugin(plugin_ref)
 
-    # TODO iterate pipelines and save each, also set this connector's profile (reuse `pipelineInFocusIndex`?)
-
     settings = PluginSettingsService(
         project, plugin, config_service=config_service, show_hidden=False
     )
 
-    for profile in payload:
-        # select the correct profile
-        name = profile["name"]
-        plugin.use_profile(plugin.get_profile(name))
+    config = payload.get("config", {})
+    for name, value in config.items():
+        if not validate_plugin_config(plugin, name, value, project, settings):
+            continue
 
-        for name, value in profile["config"].items():
-            if not validate_plugin_config(plugin, name, value, project, settings):
-                continue
+        if value == "":
+            settings.unset(name, session=db.session)
+        else:
+            settings.set(name, value, session=db.session)
 
-            if value == "":
-                settings.unset(name, session=db.session)
-            else:
-                settings.set(name, value, session=db.session)
-
-    profiles = settings.profiles_with_config(redacted=True, session=db.session)
-    for profile in profiles:
-        decorate_profile_config(profile)
-
-    return jsonify(profiles)
+    return jsonify(get_config_with_metadata(settings))
 
 
 @orchestrationsBP.route("/<plugin_ref:plugin_ref>/configuration/test", methods=["POST"])
@@ -412,9 +382,6 @@ def test_plugin_configuration(plugin_ref) -> Response:
     payload = request.get_json()
     config_service = ConfigService(project)
     plugin = config_service.get_plugin(plugin_ref)
-
-    # load the correct profile
-    plugin.use_profile(plugin.get_profile(payload.get("profile")))
 
     settings = PluginSettingsService(
         project, plugin, config_service=config_service, show_hidden=False